Mesa: [windows] Fix for RDP / Chrome Remote Desktop.
This fixesMesa to work properly under Remote Terminal or
Chrome Remote Desktop.

The main issue was that the Mesa WGL backend was still relying
on host GDI32-provided GetPixelFormat() / SetPixelFormat() functions
even though this is not needed. The patch removes this dependency.

For context see: https://android-review.googlesource.com/#/c/152706/

Change-Id: I1d5a027fdb02c492f26bbe1aa94ac4fd3a77a9bd